requirements.txt
flask==3.0.3
google-api-python-client==2.153.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
requests==2.32.3


